requirements.txt
python-telegram-bot[job-queue]>=21.0
fastapi>=0.115.0
uvicorn>=0.34.0
aiosqlite>=0.20.0
pydantic>=2.0
python-dotenv>=1.0
httpx>=0.27.0
pytest>=8.0
pytest-asyncio>=0.24.0
